+++ src/main/java/com/example/exchangeapp/dto/CommissionDto.java
@@ -5,22 +5,23 @@ import com.example.exchangeapp.enums.Currency;
 import javax.validation.constraints.DecimalMax;
 import javax.validation.constraints.DecimalMin;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 public class CommissionDto {
 
     @DecimalMin("0.0")
     @DecimalMax("100.0")
-    private double commissionPt;
+    private BigDecimal commissionPt;
     @NotNull
     private Currency from;
     @NotNull
     private Currency to;
 
-    public double getCommissionPt() {
+    public BigDecimal getCommissionPt() {
         return commissionPt;
     }
 
-    public void setCommissionPt(double commissionPt) {
+    public void setCommissionPt(BigDecimal commissionPt) {
         this.commissionPt = commissionPt;
     }
